WIP: move try it into own workspace
