+++ packages/api-explorer/src/components/index.ts
@@ -9,4 +9,3 @@ export { DocSDKs } from './DocSDKs'
 export { Main } from './MainWrapper'
 export { SideNav } from './SideNav'
 export { SideNavToggle } from './SideNavToggle'
-export { TryIt, TryItInput, TryItCallback, TryItHttpMethod } from './TryIt'
