+++ packages/api-explorer/src/scenes/MethodScene/MethodScene.tsx
@@ -24,8 +24,8 @@
 
 import React, { FC } from 'react'
 import { Flex, Space } from '@looker/components'
-import { ApiModel, typeRefs } from '@looker/sdk-codegen'
 import { useParams } from 'react-router-dom'
+import { TryIt, TryItHttpMethod } from '@looker/try-it'
 
 import {
   DocMarkdown,
@@ -33,9 +33,8 @@ import {
   DocTitle,
   DocSDKs,
   Main,
-  TryIt,
-  TryItHttpMethod,
 } from '../../components'
+
 import {
   DocActivityType,
   DocRateLimited,
@@ -44,6 +43,7 @@ import {
   DocOperation,
 } from './components'
 import { createInputs } from './utils'
+import { ApiModel, typeRefs } from '@looker/sdk-codegen'
 
 interface DocMethodProps {
   api: ApiModel
