+++ packages/api-explorer/src/scenes/MethodScene/utils.ts
@@ -32,7 +32,7 @@ import {
   IType,
   IMethod,
 } from '@looker/sdk-codegen'
-import { TryItInput } from '../../components'
+import { TryItInput } from '../../../../try-it'
 
 // TODO: use potential equivalent from sdk-codegen, confirm formats
 export const getTypeDefault = (type: string) => {
