+++ packages/try-it/jest.setup.js
@@ -1,0 +1,2 @@
+require('@testing-library/jest-dom/extend-expect')
+require('jest-styled-components')
