+++ packages/try-it/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@looker/api-explorer",
+  "name": "@looker/try-it",
   "version": "0.1.0-alpha",
-  "description": "Looker API Explorer",
+  "description": "A dynamic REST request test form",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "files": [
@@ -24,71 +24,35 @@
     "test": "jest -- src/"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.9.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
-    "@babel/plugin-transform-runtime": "^7.9.0",
-    "@babel/preset-env": "^7.9.0",
-    "@babel/preset-react": "^7.9.4",
-    "@babel/preset-typescript": "^7.9.0",
     "@looker/components-test-utils": "^0.7.36",
-    "@looker/eslint-config": "^1.0.14",
-    "@looker/prettier-config": "^1.0.15",
     "@testing-library/jest-dom": "^5.8.0",
     "@testing-library/react": "^10.0.4",
     "@testing-library/user-event": "^10.4.0",
     "@types/jest": "^25.2.3",
     "@types/lodash": "^4.14.153",
-    "@types/react-dom": "^16.9.6",
-    "@types/react-router-dom": "^5.1.3",
     "@types/react-test-renderer": "^16.9.2",
     "@types/styled-components": "^4.4.1",
     "@types/styled-system": "^5.1.9",
     "ace": "^1.3.0",
     "ace-builds": "^1.4.11",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-styled-components": "^1.10.7",
-    "css-loader": "^3.4.2",
     "enzyme": "^3.11.0",
-    "history": "^4.10.1",
     "jest": "^26.0.1",
     "jest-config": "^25.3.0",
     "jest-styled-components": "^6.3.4",
     "react-test-renderer": "^16.13.1",
     "style-loader": "^1.1.3",
     "ts-jest": "^26.0.0",
-    "typescript": "3.8.2",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
-  },
-  "prettier": "@looker/prettier-config",
-  "stylelint": {
-    "extends": "@looker/eslint-config/stylelint"
-  },
-  "eslintConfig": {
-    "extends": "@looker/eslint-config",
-    "rules": {
-      "sort-keys": "off",
-      "@typescript-eslint/camelcase": "off"
-    }
+    "typescript": "3.8.2"
   },
   "dependencies": {
-    "@babel/runtime": "^7.9.2",
     "@looker/components": "^0.8.6",
     "@looker/sdk": "^0.3.3-alpha.0",
     "@looker/sdk-codegen": "^0.1.2-alpha.18+551e4fe",
     "@looker/sdk-codegen-utils": "^0.1.0-alpha.2",
-    "@looker/try-it": "*",
     "@testing-library/jest-dom": "^5.5.0",
     "@types/lodash": "^4.14.144",
     "@types/node": "^13.13.4",
     "@types/readable-stream": "^2.3.5",
-    "eslint-import-resolver-webpack": "^0.12.1",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-prettier": "^3.1.3",
     "lerna": "^3.20.2",
     "lint-staged": "^10.2.2",
     "lodash": "^4.17.15",
@@ -97,9 +61,6 @@
     "react-ace": "^8.1.0",
     "react-dom": "^16.13.1",
     "react-is": "^16.13.1",
-    "react-markdown": "^4.3.1",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
     "styled-components": "^4.4.1",
     "typescript": "^3.8.2"
   },
